TESS/requirements.txt
alabaster==0.7.10
Babel==2.4.0
certifi==2017.7.27.1
chardet==3.0.4
colorama==0.3.9
docutils==0.14
idna==2.6
imagesize==0.7.1
influxdb==3.0.0
Jinja2==2.9.6
MarkupSafe==1.0
pydbus==0.6.0
Pygments==2.2.0
PyScaffold==2.5.7
python-dateutil==2.5.3
pytz==2016.6.1
PyYAML==3.12
requests==2.11.1
schedule==0.3.2
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.6.3
sphinxcontrib-websupport==1.0.1
typing==3.6.2
urllib3==1.22
